Reutilizacion boton atras, reutilizacion assets  fondo del boton.
