--- src/pages/AdminControlPanel/AddEventPage.jsx
@@ -1,31 +1,0 @@
-import React, { useState } from 'react';
-import { Snackbar, Alert, Container } from '@mui/material';
-import FormAddEvent from './components/FormAddEvent'; // Ajusta la ruta según tu estructura
-import Header from '../../components/HeaderPublic';
-
-
-const AddEventPage = () => {
-    const [events, setEvents] = useState([]);
-    const [success, setSuccess] = useState(false);
-
-    const handleAddEvent = (event) => {
-        setEvents([...events, { ...event, id: Date.now() }]);
-        setSuccess(true);
-    };
-
-    return (
-        <>
-            <Header /> {/* Usa el Header */}
-            <Container maxWidth="md" sx={{ mt: 4 }}>
-                <FormAddEvent onSubmit={handleAddEvent} />
-                <Snackbar open={success} autoHideDuration={6000} onClose={() => setSuccess(false)}>
-                    <Alert onClose={() => setSuccess(false)} severity="success">
-                        ¡Evento agregado con éxito!
-                    </Alert>
-                </Snackbar>
-            </Container>
-        </>
-    );
-};
-
-export default AddEventPage;
