+++ src/pages/GestionEscultores/components/Escultores/CrearEscultores.jsx
@@ -3,9 +3,10 @@ import { Box, TextField, Typography, Button, Icon } from '@mui/material';
 import ImageIcon from '@mui/icons-material/Image';
 import HeaderPublic from '../../../../components/HeaderPublic';
 import Footer from '../../../../components/Footer';
-import fondoBoton from '../../../../assets/gestioneventos/Rectangle 32.svg';
+import fondoBoton from '../../../../assets/fondobutton/Rectangle 32.svg';
 import { useNavigate } from 'react-router-dom';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+import BackButton from '../../../../components/BackButton';
 // import { createEscultor } from '../../../../api/escultores.routes'; // Importar la función desde escultores.routes.js
 
 const CrearEscultor = () => {
@@ -214,15 +215,7 @@ const CrearEscultor = () => {
                         >
                             Crear Escultor
                         </Button>
-                        <Button
-                            startIcon={<ArrowBackIcon />}
-                            variant="outlined"
-                            color="secondary"
-                            onClick={handleAtras}
-                            sx={{ width: '48%' }}
-                        >
-                            Atrás
-                        </Button>
+                        <BackButton sx={{ width: '48%' }} />
                     </Box>
                 </Box>
             </Box>
