+++ src/pages/GestionEsculturas/components/EsculturaButton.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, Box } from '@mui/material';
 import AddIcon from '@mui/icons-material/Add';
-import fondoBoton from '../../../assets/gestionesculturas/Rectangle 32.svg';
+import fondoBoton from '../../../assets/fondobutton/Rectangle 32.svg';
 import { useNavigate } from 'react-router-dom';
 
 const EsculturaButton = () => {
